pyproject.toml
[project]
name = "oversearegister"
version = "0.1.0"
description = ""
authors = [
    {name = "Anthony Sung",email = "sungpinyue@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "reportlab (>=4.4.3,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
